update shared files structure
