+++ renderer/src/components/ActivitiesSection/ActivitiesSection.tsx
@@ -7,8 +7,8 @@ import { Square2StackIcon } from "@heroicons/react/24/outline";
 import { loadGoogleEventsFromAllUsers } from "../../helpers/utils/google";
 import { getOffice365Events } from "../../helpers/utils/office365";
 import { checkIsToday } from "../../helpers/utils/datetime-ui";
-import ButtonTransparent from "../../shared/ButtonTransparent";
-import Popup from "../../shared/Popup";
+import { ButtonTransparent } from "../../shared/ButtonTransparent";
+import Popup from "../../shared/Popup/Popup";
 import { useMainStore } from "../../store/mainStore";
 import { shallow } from "zustand/shallow";
 import { ActivitiesSectionProps, PlaceholderProps } from './types';
