+++ renderer/src/components/UpdateDescription/UpdateDescription.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from "react";
 import { shallow } from "zustand/shallow";
 import { useUpdateStore } from "../../store/updateStore";
-import DisclosureSection from "../../shared/DisclosureSection";
+import { DisclosureSection } from "../../shared/DisclosureSection";
 import { GlobeAltIcon } from "@heroicons/react/24/solid";
-import SlackIcon from "../../shared/SlackIcon";
+import SlackIcon from "../../shared/SlackIcon/SlackIcon";
 import { Release } from './types';
 
 export default function UpdateDescription() {
