+++ renderer/src/shared/AutocompleteSelector/index.ts
@@ -1,0 +1,1 @@
+export { default as AutocompleteSelector } from './AutocompleteSelector';
