+++ renderer/src/shared/BetaToggle/index.ts
@@ -1,0 +1,1 @@
+export { default as BetaToggle } from './BetaToggle';
