+++ renderer/src/shared/Button/Button.tsx
@@ -1,25 +1,13 @@
-import Loader from "./Loader";
-import { ReactNode } from "react";
-
-type ButtonProps = {
-  text: string;
-  callback?: () => void;
-  disabled?: boolean;
-  status?: string;
-  type?: "button" | "submit" | "reset";
-  tabIndex?: number;
-  children?: ReactNode;
-};
-
-const basicStyles =
-  "inline-flex items-center justify-center px-4 py-2 gap-2 text-sm font-medium border rounded-md shadow-sm dark:border-dark-form-back ";
-const defaultStyles = "bg-blue-600 text-white dark:bg-dark-button-back ";
-const saveHoverStyles = "hover:bg-blue-700 hover:dark:bg-dark-button-hover ";
-const cancelHoverStyles = "hover:bg-gray-50 ";
-const focusStyles =
-  "focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 focus:dark:border-focus-border focus:dark:ring-focus-border ";
-const disabledStyles =
-  "disabled:pointer-events-none disabled:bg-gray-300 disabled:dark:bg-gray-600 disabled:dark:border-dark-border ";
+import { Loader } from "../Loader";
+import { ButtonProps } from './types';
+import {
+  basicStyles,
+  defaultStyles,
+  saveHoverStyles,
+  cancelHoverStyles,
+  focusStyles,
+  disabledStyles
+} from './constants';
 
 export default function Button({
   callback,
