+++ renderer/src/shared/Button/constants.ts
@@ -1,0 +1,9 @@
+export const basicStyles =
+  "inline-flex items-center justify-center px-4 py-2 gap-2 text-sm font-medium border rounded-md shadow-sm dark:border-dark-form-back ";
+export const defaultStyles = "bg-blue-600 text-white dark:bg-dark-button-back ";
+export const saveHoverStyles = "hover:bg-blue-700 hover:dark:bg-dark-button-hover ";
+export const cancelHoverStyles = "hover:bg-gray-50 ";
+export const focusStyles =
+  "focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 focus:dark:border-focus-border focus:dark:ring-focus-border ";
+export const disabledStyles =
+  "disabled:pointer-events-none disabled:bg-gray-300 disabled:dark:bg-gray-600 disabled:dark:border-dark-border ";
