+++ renderer/src/shared/Button/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+export type ButtonProps = {
+  text: string;
+  callback?: () => void;
+  disabled?: boolean;
+  status?: string;
+  type?: "button" | "submit" | "reset";
+  tabIndex?: number;
+  children?: ReactNode;
+};
