+++ renderer/src/shared/ButtonTransparent/index.ts
@@ -1,0 +1,1 @@
+export { default as ButtonTransparent } from './ButtonTransparent';
