+++ renderer/src/shared/ButtonTransparent/types.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from "react";
+
+export type ButtonTransparentProps = {
+  children: ReactNode;
+  callback?: () => void;
+};
