+++ renderer/src/shared/DeleteMessage/index.ts
@@ -1,0 +1,1 @@
+export { default as DeleteMessage } from './DeleteMessage';
