+++ renderer/src/shared/DeleteMessage/types.ts
@@ -1,0 +1,6 @@
+export type DeleteMessageProps = {
+  setShowDeleteButton: (value: boolean) => void;
+  setShowDeleteMessage: (value: boolean) => void;
+  selectedDate: Date;
+  setSelectedDateReport: (value: string) => void;
+};
