+++ renderer/src/shared/DisclosureSection/index.ts
@@ -1,0 +1,1 @@
+export { default as DisclosureSection } from './DisclosureSection';
