+++ renderer/src/shared/DisclosureSection/types.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from "react";
+
+export type DisclosureSectionProps = {
+  toggleFunction: () => void;
+  isOpen: boolean;
+  title: string;
+  children: ReactNode;
+};
