+++ renderer/src/shared/ErrorPlaceholder/index.ts
@@ -1,0 +1,2 @@
+export { default as ErrorPlaceholder } from './ErrorPlaceholder';
+export * from './types';
