+++ renderer/src/shared/ErrorPlaceholder/types.ts
@@ -1,0 +1,4 @@
+export type RenderError = {
+  errorTitle: string;
+  errorMessage: string;
+};
