+++ renderer/src/shared/MenuItem/index.ts
@@ -1,0 +1,1 @@
+export { default as MenuItem } from './MenuItem';
