+++ renderer/src/shared/MenuItem/types.ts
@@ -1,0 +1,7 @@
+import { ReactNode } from "react";
+
+export type MenuItemProps = {
+  children: ReactNode;
+  callback?: () => void;
+  isActive?: boolean;
+};
