+++ renderer/src/shared/NavButtons/index.ts
@@ -1,0 +1,1 @@
+export { default as NavButtons } from './NavButtons';
