+++ renderer/src/shared/NavButtons/types.ts
@@ -1,0 +1,4 @@
+export type NavButtonsProps = {
+  prevCallback: () => void;
+  nextCallback: () => void;
+};
