+++ renderer/src/shared/Popup/Popup.tsx
@@ -1,23 +1,5 @@
-const buttonColors = {
-  gray: "bg-white ring-1 ring-gray-300 text-gray-900 hover:bg-gray-50 dark:ring-gray-700 dark:text-dark-heading dark:bg-gray-500 hover:dark:bg-gray-400",
-  green:
-    "bg-green-600 hover:bg-green-500 dark:text-dark-heading dark:bg-green-600/70 hover:dark:bg-green-500/70",
-  red: "bg-red-600 hover:bg-red-500 dark:text-dark-heading dark:bg-red-600/70 hover:dark:bg-red-500/70",
-};
-
-type PopupButton = {
-  text: string;
-  color: keyof typeof buttonColors;
-  callback: () => void;
-};
-
-type PopupProps = {
-  title: string;
-  description: string;
-  top?: string;
-  left?: string;
-  buttons: PopupButton[];
-};
+import { PopupProps } from './types';
+import { buttonColors } from './constants';
 
 export default function Popup({
   title,
