+++ renderer/src/shared/Popup/constants.ts
@@ -1,0 +1,6 @@
+export const buttonColors = {
+  gray: "bg-white ring-1 ring-gray-300 text-gray-900 hover:bg-gray-50 dark:ring-gray-700 dark:text-dark-heading dark:bg-gray-500 hover:dark:bg-gray-400",
+  green:
+    "bg-green-600 hover:bg-green-500 dark:text-dark-heading dark:bg-green-600/70 hover:dark:bg-green-500/70",
+  red: "bg-red-600 hover:bg-red-500 dark:text-dark-heading dark:bg-red-600/70 hover:dark:bg-red-500/70",
+};
