+++ renderer/src/shared/Popup/index.ts
@@ -1,0 +1,1 @@
+export { default as Popup } from './Popup';
