+++ renderer/src/shared/SlackIcon/SlackIcon.tsx
