+++ renderer/src/shared/TimeBadge/__tests__/TimeBadge.test.tsx
@@ -1,4 +1,5 @@
-import TimeBadge, { TimeBadgeProps } from "./TimeBadge";
+import TimeBadge from "../TimeBadge";
+import { TimeBadgeProps } from '../types';
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
