+++ renderer/src/shared/TimeBadge/index.ts
@@ -1,0 +1,1 @@
+export { default as TimeBadge } from './TimeBadge';
