+++ renderer/src/shared/TimeBadge/types.ts
@@ -1,0 +1,5 @@
+export type TimeBadgeProps = {
+  hours: number;
+  startTime: string;
+  selectedDate: Date;
+};
