+++ renderer/src/shared/VersionMessage/index.ts
@@ -1,0 +1,1 @@
+export { default as VersionMessage } from './VersionMessages';
